requirements.txt
asgiref==3.2.3
Django==3.0.1
djangorestframework==3.11.0
Pillow==6.2.1
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
